requirements.txt
# Core dependencies
numpy>=1.21.0
torch>=1.12.0
awkward>=2.0.0

# HEP data formats
uproot>=5.0.0
tables>=3.7.0

# ML frameworks
ray[default]>=2.0.0
ray[train]>=2.0.0
ray[serve]>=2.0.0
mlflow>=2.0.0
tensorboard>=2.10.0

# Data processing
pandas>=1.3.0
scikit-learn>=1.0.0
pyyaml>=6.0

# Web services
fastapi>=0.100.0
uvicorn>=0.20.0

# CLI
typer>=0.9.0
typing-extensions>=4.0.0

# ONNX
onnx>=1.12.0
onnxruntime>=1.12.0

# Documentation
mkdocs>=1.4.0
mkdocstrings[python]>=0.18.0
griffe<0.40

# Styling
black>=24.0.0
flake8>=6.0.0
isort>=5.12.0
pyupgrade>=3.3.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-xdist>=3.0.0

# Development
pre-commit>=3.2.0

